move migration logic to 'migration' package
